rename tschorle to core
