+++ lua/core/lazy.lua
